Rename Elasticity namespace and add tag for elastic stress
