+++ src/Elliptic/Systems/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+add_subdirectory(Elasticity)
 add_subdirectory(Poisson)
-add_subdirectory(ThermalNoise)
 add_subdirectory(Xcts)
