+++ src/Elliptic/Systems/Elasticity/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY ThermalNoise)
+set(LIBRARY Elasticity)
