+++ tests/Unit/Elliptic/Systems/Elasticity/CMakeLists.txt
@@ -1,15 +1,15 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY "Test_ThermalNoise")
+set(LIBRARY "Test_Elasticity")
 
 set(LIBRARY_SOURCES
-  Test_ThermalNoiseTags.cpp
+  Test_Tags.cpp
   )
 
 add_test_library(
   ${LIBRARY}
-  "Elliptic/Systems/ThermalNoise/"
+  "Elliptic/Systems/Elasticity/"
   "${LIBRARY_SOURCES}"
   ""
   )
